Add and remove some packages and improve script

+++ setup.sh
@@ -31,7 +31,7 @@ check() {
 }
 
 warning() {
-  echo "$warning>⚠️  $1"
+  echo "$warning> ⚠️  $1"
 }
 
 install_packages() {
@@ -43,12 +43,6 @@ install_packages() {
 update_system() {
   step "Updating system"
   sudo apt update && sudo apt full-upgrade -y
-  check
-}
-
-remove_packages() {
-  step "Removing APT packages"
-  sudo apt purge -y apport
   check
 }
 
@@ -65,127 +59,120 @@ create_folders() {
 }
 
 install_brave() {
-  step "Installing Brave Browser"
-
   if exists brave-browser; then
-    warning "Docker is already installed, skipping install"
-  else
-    sudo apt install -y apt-transport-https curl
-    sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
-    echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg arch=amd64] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
-    sudo apt update
-    sudo apt install -y brave-browser
+    warning "Brave Browser already installed, skipping"
+    return
   fi
 
+  step "Installing Brave Browser"
+  sudo curl -fsSLo /usr/share/keyrings/brave-browser-archive-keyring.gpg https://brave-browser-apt-release.s3.brave.com/brave-browser-archive-keyring.gpg
+  echo "deb [signed-by=/usr/share/keyrings/brave-browser-archive-keyring.gpg] https://brave-browser-apt-release.s3.brave.com/ stable main"|sudo tee /etc/apt/sources.list.d/brave-browser-release.list
+  sudo sh -c "apt update; apt install brave-browser -y"
   check
 }
 
 install_nvm() {
-  step "Installing NVM"
-
   if exists nvm; then
-    warning "NVM is already installed, skipping install"
-  else
-    curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.5/install.sh | bash
-    export NVM_DIR="$HOME/.nvm"
-    [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
-    nvm install stable
-    nvm use stable
-    nvm alias default stable
+    warning "NVM already installed, skipping"
+    return
   fi
 
+  step "Installing NVM"
+  curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.5/install.sh | bash
+  export NVM_DIR="$HOME/.nvm"
+  [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
+  nvm install stable
+  nvm use stable
+  nvm alias default stable
   check
 }
 
 install_neovim() {
   step "Installing Neovim"
   if exists nvim; then
-    warning "Neovim is already installed, skipping install"
-  else
-    wget https://github.com/neovim/neovim/releases/download/stable/nvim.appimage -O ~/bin/nvim
+    warning "Neovim already installed, skipping"
+    return
   fi
+  wget https://github.com/neovim/neovim/releases/download/stable/nvim.appimage -O ~/bin/nvim
   check
 }
 
 install_docker() {
-  step "Installing Docker"
-
   if exists docker; then
-    warning "Docker is already installed, skipping install"
-  else
-    curl -fsSL https://get.docker.com -o get-docker.sh
-    sh get-docker.sh
-    rm get-docker.sh
-    sudo usermod -aG docker $USER
-    docker --version
+    warning "Docker already installed, skipping"
+    return
   fi
 
+  step "Installing Docker"
+  curl -fsSL https://get.docker.com -o get-docker.sh
+  sh get-docker.sh
+  rm get-docker.sh
+  sudo usermod -aG docker $USER
+  docker --version
   check
 }
 
 install_docker_compose() {
-  step "Installing Docker Compose"
-
   if exists docker-compose; then
-    warning "Docker-compose is already installed, skipping install"
-  else
-    sudo curl -L "https://github.com/docker/compose/releases/download/$(get_latest_release docker/compose)/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-    sudo chmod +x /usr/local/bin/docker-compose
-    docker-compose --version
+    warning "Docker Compose already installed, skipping"
+    return
   fi
 
+  step "Installing Docker Compose"
+  sudo curl -L "https://github.com/docker/compose/releases/download/$(get_latest_release docker/compose)/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+  sudo chmod +x /usr/local/bin/docker-compose
+  docker-compose --version
   check
 }
 
 install_stacer() {
-  step "Installing Stacer"
-
   if exists stacer; then
-    warning "Stacer is already installed, skipping install"
-  else
-    sudo apt install -y stacer
+    warning "Stacer already installed, skipping"
+    return
   fi
 
-  check
-}
-
-install_telegram() {
-  step "Installing Telegram"
-  sudo add-apt-repository -y ppa:atareao/telegram
-  sudo apt-get install -y telegram
+  step "Installing Stacer"
+  sudo apt install -y stacer
   check
 }
 
 install_spotify() {
+  if exists spotify-client; then
+    warning "Spotify already installed, skipping"
+    return
+  fi
+
   step "Installing Spotify"
-  curl -sS https://download.spotify.com/debian/pubkey_5E3C45D7B312C643.gpg | sudo apt-key add -
+  curl -sS https://download.spotify.com/debian/pubkey_6224F9941A8AA6D1.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
   echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-  sudo apt-get update
-  sudo apt-get install -y spotify-client
+  sudo sh -c "apt-get update ; apt-get install -y spotify-client"
   check
 }
 
 install_ghcli() {
-  step "Installing GH-CLI"
-  curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo gpg --dearmor -o /usr/share/keyrings/githubcli-archive-keyring.gpg
-  echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null
-  sudo apt update
-  sudo apt install -y gh
-  #gh auth login
-  check
-}
+  if exists gh; then
+    warning "GH-CLI already installed, skipping"
+    return
+  fi
 
-install_gcloud_sdk() {
-  step "Installing gcloud sdk"
-  echo "deb [signed-by=/usr/share/keyrings/cloud.google.gpg] https://packages.cloud.google.com/apt cloud-sdk main" | sudo tee -a /etc/apt/sources.list.d/google-cloud-sdk.list
-  sudo apt install -y apt-transport-https ca-certificates gnupg
-  curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo apt-key --keyring /usr/share/keyrings/cloud.google.gpg add -
-  sudo apt-get update && sudo apt-get install google-cloud-sdk
+  step "Installing GH-CLI"
+  (type -p wget >/dev/null || (sudo apt update && sudo apt-get install wget -y)) \
+    && sudo mkdir -p -m 755 /etc/apt/keyrings \
+    && wget -qO- https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo tee /etc/apt/keyrings/githubcli-archive-keyring.gpg > /dev/null \
+    && sudo chmod go+r /etc/apt/keyrings/githubcli-archive-keyring.gpg \
+    && echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list > /dev/null \
+    && sudo apt update \
+    && sudo apt install gh -y
   check
 }
 
 install_bitwarden() {
-  step "Installing bitwarden"
+  if exists bitwarden; then
+    warning "Bitwarden already installed, skipping"
+    return
+  fi
+
+  step "Installing Bitwarden"
   wget "https://vault.bitwarden.com/download/?app=desktop&platform=linux&variant=deb" -O bitwarden.deb
   sudo dpkg -i bitwarden.deb
   rm bitwarden.deb
@@ -193,10 +180,15 @@ install_bitwarden() {
 }
 
 install_hashicorp() {
+  if exists terraform; then
+    warning "Terraform already installed, skipping"
+    return
+  fi
+
   step "Installing hashicorp stuff"
-  curl -fsSL https://apt.releases.hashicorp.com/gpg | sudo apt-key add -
-  sudo apt-add-repository -y "deb [arch=amd64] https://apt.releases.hashicorp.com $(lsb_release -cs) main"
-  sudo apt-get update && sudo apt install -y vagrant terraform
+  wget -O- https://apt.releases.hashicorp.com/gpg | gpg --dearmor | sudo tee /usr/share/keyrings/hashicorp-archive-keyring.gpg
+  echo "deb [signed-by=/usr/share/keyrings/hashicorp-archive-keyring.gpg] https://apt.releases.hashicorp.com $(lsb_release -cs) main" | sudo tee /etc/apt/sources.list.d/hashicorp.list
+  sudo apt-get update && sudo apt install -y terraform
   check
 }
 
@@ -218,6 +210,11 @@ install_pip_packages() {
 
 install_fonts() {
   step "Installing FantasqueSansMono nerd font"
+  if [ -f "$HOME/.fonts/FantasqueSansMono.ttf" ]; then
+    warning "Font already installed, skipping"
+    return
+  fi
+
   mkdir -p $HOME/.fonts/
   wget https://github.com/ryanoasis/nerd-fonts/releases/download/v2.1.0/FantasqueSansMono.zip
   unzip FantasqueSansMono.zip
@@ -228,6 +225,11 @@ install_fonts() {
 
 install_themes() {
   step "Installing dracula gedit theme"
+  if [ -f "$HOME/.local/share/gedit/styles/dracula.xml" ]; then
+    warning "Dracula theme already installed, skipping"
+    return
+  fi
+
   mkdir -p $HOME/.local/share/gedit/styles/
   wget https://raw.githubusercontent.com/dracula/gedit/master/dracula.xml
   mv dracula.xml $HOME/.local/share/gedit/styles/
@@ -235,19 +237,25 @@ install_themes() {
 }
 
 install_starship() {
+  if exists starship; then
+    warning "Starship already installed, skipping"
+    return
+  fi
+
   step "Installing starship"
   curl -sS https://starship.rs/install.sh | sh
   check
 }
 
 install_fzf() {
-  step "Installing fzf"
-  if exists fzf; then
-    warning "FZF is already installed, skipping install"
-  else
-    git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf;
-    ~/.fzf/install
+  if [ -d "$HOME/.fzf" ]; then
+    warning "fzf already installed, skipping"
+    return
   fi
+
+  step "Installing fzf"
+  git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
+  ~/.fzf/install
   check
 }
 
@@ -295,17 +303,26 @@ tweak_inotify() {
   check
 }
 
+add_to_sudoers() {
+  step "Adding user to sudoers group"
+  if groups $USER | grep -q "\bsudo\b"; then
+    warning "User already in sudo group, skipping"
+  else
+    sudo usermod -aG sudo $USER
+    check
+  fi
+}
+
 setup() {
-  echo "\n Verzola's Ubuntu 22.04 Setup"
+  echo "\n Verzola's Ubuntu Setup"
 
   install_packages
   update_system
-  remove_packages
   cleanup_packages
   create_folders
+  install_brave
   install_nvm
   install_npm_packages
-  install_pip_packages
   install_fonts
   install_themes
   install_starship
@@ -313,19 +330,16 @@ setup() {
   install_neovim
   install_docker
   install_docker_compose
-  install_brave
   install_bitwarden
   install_hashicorp
-  install_gcloud_sdk
   install_ghcli
   install_stacer
-  install_telegram
   install_spotify
   adjust_clock
   configure_dotfiles
-  configure_tmux
   configure_zsh
   tweak_inotify
+  add_to_sudoers
 
   echo "\nFinished!"
 }
